otel exporters package refactored and updated to support metrics export
